Silence a warning.
git-svn-id: https://svn.dealii.org/trunk@16537 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/examples/step-31/step-31.cc
@@ -220,7 +220,7 @@ void TrilinosAmgPreconditioner<dim>::initialize (
 		++col_counter;
 	      }
 	  }
-	Assert (col_counter == row_lengths[row],
+	Assert (col_counter == static_cast<unsigned int>(row_lengths[row]),
 		ExcMessage("Filtering out zeros could not "
 			    "be successfully finished!"));
